+++ src/main/java/com/utn/TPFUDEE/Services/MeasurementService.java
@@ -58,8 +58,11 @@ public class MeasurementService {
     }
 
     public void deleteById(Integer id){
-        this.getById(id);
-        measurementRepository.deleteById(id);
+        if(this.getById(id)!=null)
+            measurementRepository.deleteById(id);
+        else
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Measurement Not Found For Delete");
+
     }
 
     public MoneyAndKwProjection getAddressConsumes(Integer id, String from, String to){
