+++ src/main/java/com/utn/TPFUDEE/Services/UserService.java
@@ -44,8 +44,11 @@ public class UserService {
     }
 
     public void deleteById(Integer id){
-        this.getById(id);
-        userRepository.deleteById(id);
+        if(this.getById(id)!=null)
+            userRepository.deleteById(id);
+        else
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User Not Found For Delete");
+
     }
 
     public User getUserByUserNameAndPassword(String userName, String password) {
